server/myanimearchive/requirements.txt
Flask
celery[redis]
Flask-Cors==3.0.10
pymongo
dnspython==2.1.0
python-decouple==3.5